# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Jewellery management API — retail, wholesale, GST, Girvi, ledger, reports
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /karigars:
    get:
      operationId: listKarigars
      tags: [karigar]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Karigar" }
    post:
      operationId: createKarigar
      tags: [karigar]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/KarigarInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Karigar" }
  /karigars/{id}:
    put:
      operationId: updateKarigar
      tags: [karigar]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/KarigarInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Karigar" }
    delete:
      operationId: deleteKarigar
      tags: [karigar]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }
  /karigar-jobs:
    get:
      operationId: listKarigarJobs
      tags: [karigar]
      parameters:
        - in: query
          name: status
          schema: { type: string, enum: [issued, received, all] }
        - in: query
          name: karigarId
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/KarigarJob" }
    post:
      operationId: createKarigarJob
      tags: [karigar]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/KarigarJobInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/KarigarJob" }
  /karigar-jobs/{id}/receive:
    post:
      operationId: receiveKarigarJob
      tags: [karigar]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/KarigarReceiveInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/KarigarJob" }
  /karigar-jobs/{id}:
    delete:
      operationId: deleteKarigarJob
      tags: [karigar]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }
  /repairs:
    get:
      operationId: listRepairs
      tags: [repairs]
      parameters:
        - in: query
          name: status
          schema: { type: string, enum: [received, in_progress, ready, delivered, all] }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/RepairJob" }
    post:
      operationId: createRepair
      tags: [repairs]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RepairInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RepairJob" }
  /repairs/{id}:
    put:
      operationId: updateRepair
      tags: [repairs]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RepairUpdateInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RepairJob" }
    delete:
      operationId: deleteRepair
      tags: [repairs]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }
...
components:
  schemas:
    Ok:
      type: object
      properties:
        ok: { type: boolean }
      required: [ok]
...
    Karigar:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        phone: { type: string }
        speciality: { type: string, nullable: true }
        address: { type: string, nullable: true }
        notes: { type: string, nullable: true }
        activeJobs: { type: integer }
        totalJobs: { type: integer }
        createdAt: { type: string, format: date-time }
      required: [id, name, phone, activeJobs, totalJobs, createdAt]
    KarigarInput:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        phone: { type: string, minLength: 1 }
        speciality: { type: string, nullable: true }
        address: { type: string, nullable: true }
        notes: { type: string, nullable: true }
      required: [name, phone]
    KarigarJob:
      type: object
      properties:
        id: { type: string }
        jobNumber: { type: string }
        karigarId: { type: string }
        karigarName: { type: string }
        itemDescription: { type: string }
        metal: { type: string }
        purity: { type: string }
        issuedWeight: { type: number }
        receivedWeight: { type: number, nullable: true }
        wastageWeight: { type: number, nullable: true }
        expectedWastagePct: { type: number }
        actualWastagePct: { type: number, nullable: true }
        laborCharge: { type: number }
        issuedDate: { type: string, format: date-time }
        expectedDate: { type: string, format: date-time, nullable: true }
        receivedDate: { type: string, format: date-time, nullable: true }
        status: { type: string, enum: [issued, received] }
        notes: { type: string, nullable: true }
      required: [id, jobNumber, karigarId, karigarName, itemDescription, metal, purity, issuedWeight, expectedWastagePct, laborCharge, issuedDate, status]
    KarigarJobInput:
      type: object
      properties:
        karigarId: { type: string }
        itemDescription: { type: string, minLength: 1 }
        metal: { type: string, enum: [gold, silver, platinum] }
        purity: { type: string, minLength: 1 }
        issuedWeight: { type: number }
        expectedWastagePct: { type: number }
        laborCharge: { type: number }
        issuedDate: { type: string, format: date-time }
        expectedDate: { type: string, format: date-time, nullable: true }
        notes: { type: string, nullable: true }
      required: [karigarId, itemDescription, metal, purity, issuedWeight, expectedWastagePct, laborCharge, issuedDate]
    KarigarReceiveInput:
      type: object
      properties:
        receivedWeight: { type: number }
        receivedDate: { type: string, format: date-time }
        notes: { type: string, nullable: true }
      required: [receivedWeight, receivedDate]
    RepairJob:
      type: object
      properties:
        id: { type: string }
        ticketNumber: { type: string }
        customerId: { type: string }
        customerName: { type: string }
        itemDescription: { type: string }
        metal: { type: string }
        purity: { type: string, nullable: true }
        weightGrams: { type: number }
        issue: { type: string }
        estimatedCost: { type: number }
        finalCost: { type: number, nullable: true }
        receivedDate: { type: string, format: date-time }
        promisedDate: { type: string, format: date-time, nullable: true }
        deliveredDate: { type: string, format: date-time, nullable: true }
        status: { type: string, enum: [received, in_progress, ready, delivered] }
        paidAmount: { type: number }
        notes: { type: string, nullable: true }
      required: [id, ticketNumber, customerId, customerName, itemDescription, metal, weightGrams, issue, estimatedCost, receivedDate, status, paidAmount]
    RepairInput:
      type: object
      properties:
        customerId: { type: string }
        itemDescription: { type: string, minLength: 1 }
        metal: { type: string, enum: [gold, silver, platinum, diamond] }
        purity: { type: string, nullable: true }
        weightGrams: { type: number }
        issue: { type: string, minLength: 1 }
        estimatedCost: { type: number }
        receivedDate: { type: string, format: date-time }
        promisedDate: { type: string, format: date-time, nullable: true }
        notes: { type: string, nullable: true }
      required: [customerId, itemDescription, metal, weightGrams, issue, estimatedCost, receivedDate]
    RepairUpdateInput:
      type: object
      properties:
        status: { type: string, enum: [received, in_progress, ready, delivered] }
        finalCost: { type: number, nullable: true }
        paidAmount: { type: number, nullable: true }
        deliveredDate: { type: string, format: date-time, nullable: true }
        notes: { type: string, nullable: true }
      required: [status]
...