# lib/api-zod/src/generated/types/karigar.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Jewellery management API — retail, wholesale, GST, Girvi, ledger, reports
 * OpenAPI spec version: 0.1.0
 */

export interface Karigar {
  id: string;
  name: string;
  phone: string;
  speciality?: string | null;
  address?: string | null;
  notes?: string | null;
  activeJobs: number;
  totalJobs: number;
  createdAt: Date;
}