# lib/api-zod/src/generated/types/karigarReceiveInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Jewellery management API — retail, wholesale, GST, Girvi, ledger, reports
 * OpenAPI spec version: 0.1.0
 */

export interface KarigarReceiveInput {
  receivedWeight: number;
  receivedDate: Date;
  notes?: string | null;
}