# lib/api-zod/src/generated/types/repairUpdateInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Jewellery management API — retail, wholesale, GST, Girvi, ledger, reports
 * OpenAPI spec version: 0.1.0
 */
import type { RepairUpdateInputStatus } from "./repairUpdateInputStatus";

export interface RepairUpdateInput {
  status: RepairUpdateInputStatus;
  finalCost?: number | null;
  paidAmount?: number | null;
  deliveredDate?: Date | null;
  notes?: string | null;
}